Add more type specs to our Types
