+++ lib/graphql/type/enum_value.ex
@@ -1,0 +1,8 @@
+defmodule GraphQL.Type.EnumValue do
+  @type t :: %GraphQL.Type.EnumValue{
+    name: binary,
+    description: binary,
+    value: any
+  }
+  defstruct name: "", value: "", description: ""
+end
