+++ lib/graphql/type/input.ex
@@ -1,5 +1,9 @@
 defmodule GraphQL.Type.Input do
-  defstruct name: "Float",
-    description: "",
-    fields: %{}
+  @type t :: %GraphQL.Type.Input{
+    name: binary,
+    description: binary,
+    fields: map
+  }
+
+  defstruct name: "Input", description: "", fields: %{}
 end
