+++ lib/graphql/type/non_null.ex
@@ -1,0 +1,4 @@
+defmodule GraphQL.Type.NonNull do
+  @type t :: %{ofType: map}
+  defstruct ofType: nil
+end
